+++ CHANGELOG.md
@@ -13,7 +13,7 @@ Changes in this section will be included in the next release.
 #### Added
 - The CLI now enables the creation, listing, and getting of parameter sweeps. Parameter sweeps enable one to pass specific values to a build to, for example, search for an optimal setting for a particular component. A parameter sweep can be created like a batch, but with the addition of either:
   - A `parameter-name` and `parameter-values` flag pair that enable a single dimension sweep with a comma separated list of values for the named parameter
-  - A `grid-search-config` file can be passed, as a json list: `[{"name" : "param", "values" : ["value1",value2"]}, ...]` for example. This can create a multi-dimensional grid search.
+  - A `grid-search-config` file can be passed, as a json list: `[{"name" : "param", "values" : [ "value1", "value2" ]}, ...]` for example. This can create a multi-dimensional grid search.
 
 ### v0.1.26 - October 26 2023